data/cloud-functions/requirements.txt
requests==2.25.0
python-dateutil==2.8.1
